Added other complete code

+++ src/fuse_fs/deadlock_aware_fuse.cpp
@@ -7,8 +7,8 @@
 #include<sys/socket.h>
 #include<sys/un.h>
 #include <dirent.h>
-#include <signal.h>
 #include <fcntl.h>
+#include <signal.h>
 #include <sys/stat.h>
 #include <cstring>
 #include <cerrno>
@@ -24,7 +24,6 @@ const char *SOCKET_PATH = "/tmp/ransomware_defense.sock";
 const char *BACKING_STORE = "/tmp/backing_store";
 pid_t ML_DAEMON_PID = 0;
 
-
 extern int check_ipc_for_response();
 
 void load_daemon_pid(){
@@ -35,74 +34,69 @@ void load_daemon_pid(){
 }
 
 void load_honeyfiles() {
-    ifstream h_file("/tmp/honeyfiles.txt");
-    string line;
-    if (h_file.is_open()) {
-        while (getline(h_file, line)) {
-            HONEYFILES.push_back(line);
-        }
+  ifstream h_file("/tmp/honeyfiles.txt");
+  string line;
+  if (h_file.is_open()) {
+    while (getline(h_file, line)) {
+      HONEYFILES.push_back(line);
     }
+  }
 }
 
-int query_ml_daemon_with_timeout(const char* buf, size_t size, int timeout_ms){
+ void log_edr_alert(const string& message) {
+      ofstream log_file("/tmp/edr_alerts.log", ios_base::app);
+      if (log_file.is_open()) {
+          log_file << message << endl;
+      }
+  }
+
+
+int query_ml_daemon_with_timeout(const char* path, const char* buf, size_t size, int timeout_ms){
     int sock = socket(AF_UNIX, SOCK_STREAM, 0);
     struct sockaddr_un addr;
     addr.sun_family = AF_UNIX;
 
-
-    // auto start_time = std::chrono::steady_clock::now();
     strncpy(addr.sun_path, SOCKET_PATH, sizeof(addr.sun_path) -1);
 
     if(connect(sock, (struct sockaddr*) & addr, sizeof(addr)) == -1){
         return -1;
     }
 
-    send(sock, buf, size, 0);
+    char path_header[512] = {0};
+    strncpy(path_header, path, 511);
+    if(send(sock, path_header, 512, 0) == -1) {
+      close(sock);
+      return -1;
+    }
 
-    struct timeval tv; //timeval has two elements called tv_sec and tv_usec
+    if(send(sock, buf, size, 0) == -1) {
+      close(sock);
+      return -1;
+    }
 
+    struct timeval tv;
     tv.tv_sec = timeout_ms / 1000;
     tv.tv_usec = (timeout_ms % 1000) * 1000;
 
     setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, (const char*)&tv, sizeof tv);
-
-    char response[2] = {0};
-
+    char response[] = {0};
     int bytes_received = recv(sock, response, 1, 0);
 
     close(sock);
 
     if(bytes_received <= 0){
         cerr<<"[Critical] ML Daemon timeout!  Breaking wait to prevent deadlock."<<endl;
-
         return -1;
     }
 
     return (response[0] == '1') ? 1: 0;
-    // while(true){
-
-    //     int ml_response = check_ipc_for_response();
-
-    //     if(ml_response == 0 || ml_response == 1){
-    //         return ml_response;
-    //     }
-        
-    //     auto current_time = std::chrono::steady_clock::now();
-    //     auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(current_time - start_time).count();
-
-    //     if(elapsed > timeout_ms){
-    //         cerr<<"[Critical]! ML Daemon timeout. Aborting Circular wait to prevent deadlock."<<endl;
-    //         return -1;
-    //     }
-
-    //     usleep(1000);
-    // }
 }
 
 void translate_path(char fpath[PATH_MAX], const char* path){
     strcpy(fpath,  BACKING_STORE);
     strncat(fpath, path, PATH_MAX - strlen(BACKING_STORE) - 1);
 }
+
 
 static int ransomware_proof_getattr(const char* path, struct stat* stbuf){
     char fpath[PATH_MAX];
@@ -141,6 +135,7 @@ static int ransomware_proof_readdir(const char* path, void* buf, fuse_fill_dir_t
     return 0;
 }
 
+
 static int ransomware_proof_open(const char *path, struct fuse_file_info *fi) {
     char fpath[PATH_MAX];
     translate_path(fpath, path);
@@ -154,6 +149,7 @@ static int ransomware_proof_open(const char *path, struct fuse_file_info *fi) {
     return 0;
 }
 
+
 static int ransomware_proof_read(const char* path, char* buf, size_t size, off_t offset, struct fuse_file_info* fi){
     int res = pread(fi->fh, buf, size, offset);
     if(res == -1){
@@ -161,6 +157,7 @@ static int ransomware_proof_read(const char* path, char* buf, size_t size, off_t
     }
     return res;
 }
+
 
 static int ransomware_proof_create(const char* path, mode_t mode , struct fuse_file_info *fi){
     char fpath[PATH_MAX];
@@ -176,46 +173,49 @@ static int ransomware_proof_create(const char* path, mode_t mode , struct fuse_f
     return 0;
 }
 
+
 int ransomware_proof_truncate(const char *path, off_t size) {
-    return 0;
+  return 0;
 }
+
 
 static int ransomware_proof_write(const char* path, const char* buf, size_t size, off_t offset, struct fuse_file_info *fi) {
     pid_t calling_pid = fuse_get_context()->pid;
-
+    
     string current_path(path);
     if (find(HONEYFILES.begin(), HONEYFILES.end(), current_path) != HONEYFILES.end()) {
-
-        cerr << "\n\n[CRITICAL] TRIPWIRE TRIGGERED! Process touched a Honeyfile.\n";
-        cerr << "[MITIGATION] Bypassing ML. Instant SIGKILL to PID: " << calling_pid << "\n\n";
-        
-        kill(calling_pid, SIGKILL);
-        return -EACCES; 
+     string alert = "";
+      alert += "[HONEYFILE TRIPWIRE] Rogue process touched '" + current_path + "'!\n";
+      alert += "[MITIGATION] C++ Kernel Driver bypassed ML and executed SIGKILL.\n";
+      
+      log_edr_alert(alert); 
+      if (calling_pid > 1 && calling_pid != getpid()) {
+          kill(calling_pid, SIGKILL);
+      }
+      
+      return -EACCES;
     }
 
     if(calling_pid == ML_DAEMON_PID){
         return pwrite(fi->fh, buf, size, offset);
     }
 
+    int verdict = query_ml_daemon_with_timeout(path, buf, size, 2000);
 
-    int verdict = query_ml_daemon_with_timeout(buf, size, 500);
-  
-    if (verdict == 1) {
-        cerr << "\n\n[CRITICAL ALERT] Ransomware behavior detected!\n";
-        cerr << "[MITIGATION] Blocking write to file: " << path << "\n";
-        cerr << "[MITIGATION] Sending SIGKILL to terminate PID: " << calling_pid << "\n\n";
-        
+    if(verdict == 1){
+        cerr<<"\n\n[CRITICAL ALERT] Ransomware behavior detected!\n";
+        cerr<<"[MITIGATION] Blocking write to file:" << path << "\n";
+        cerr<<"[MITIGATION] Sending SIGKILL to terminate PID: " << calling_pid << "\n\n";
+
         int kill_status = kill(calling_pid, SIGKILL);
         
-        if (kill_status == 0) {
-            cerr << "[SUCCESS] Rogue process " << calling_pid << " successfully terminated.\n";
-        } else {
-            cerr << "[WARNING] Failed to terminate process. It may already be dead or requires root.\n";
-        }
+        if(kill_status == 0)
+          cerr << "[SUCCESS] Rogue process " << calling_pid << " successfully terminated.\n";
+        else
+          cerr << "[WARNING] Failed to terminate process. It may already be dead or requires root.\n";
 
         return -EACCES;
     }
-
 
     else if(verdict == -1) {
         cerr<<"[FAIL-SAFE] Timeout occurred. Allowing write to prevent system Freeze"<<endl;
@@ -235,7 +235,6 @@ int main(int argc, char* argv[]){
     rfs_oper.create = ransomware_proof_create;
     rfs_oper.truncate = ransomware_proof_truncate;
     rfs_oper.write = ransomware_proof_write;
-
 
     load_daemon_pid();
     load_honeyfiles();
